Further modulized the speech engine for TTS
I want the entire speech engine to be modular and easy to move around
and dynamically change. As such, I created a new class for TTS, so that
the SpeechDriver doesn't have to handle that stuff.
