+++ src/Utilities/Speech/SpeechDriver.java
@@ -1,21 +1,14 @@
 package Utilities.Speech;
 
-import com.sun.speech.freetts.Voice;
-import com.sun.speech.freetts.VoiceManager;
-
 public class SpeechDriver {
 
-	private static final String VOICENAME_kevin = "kevin";
+	private TTS tts;
 
 	public SpeechDriver() {
-
+		tts = new TTS("kevin");
 	}
 
-	public void speak(String text) {
-		Voice voice;
-		VoiceManager voiceManager = VoiceManager.getInstance();
-		voice = voiceManager.getVoice(VOICENAME_kevin);
-		voice.allocate();
-		voice.speak(text);
+	public void Speak(String m) {
+		tts.speak(m);
 	}
 }
